Allow merging bundles that have a fixed-reg def
The previous check was overly conservative: we only need to reject
bundles that have multiple uses in the same instruction.

This can only happen with an early use and a late def, so we round
the start of each range containing a fixed def up to the start of
its instruction to detect overlaps.