Only sort parallel-moves list once (#157)
* Only sort parallel-moves list once

Checking whether sources and dests overlap can use the same sort order
that's needed later, as long as the check is updated to match.

* Review comments, and defer overlap check

Now that we're sorting by `dst` from the start, the overlap check can
wait until after we remove self-moves. If the only overlaps are
self-moves we don't need the full cycle-resolution machinery.