Simplify stack-to-stack moves (#159)
* Simplify parallel stack-to-stack moves

Most of the changes here are simply refactoring.

Resolving stack-to-stack moves first fills in the scratch register for
the parallel moves, if one is needed. So we can split that into two
phases, inlining the `stack_to_stack` function which was not used
anywhere else. That lets us establish the invariant that we only
continue into the second phase if there is at least one stack-to-stack
move.

Because `compute` consumes `self`, we know that the lazily-allocated
stack-to-stack scratch register has not been allocated yet at entry to
the loop, but because the loop only runs if there are stack-to-stack
moves, we know the scratch register will definitely be allocated. So we
can remove the corresponding state from the `MoveVecWithScratch` struct
and keep it in locals with shorter names instead, eagerly initialized
before the start of the loop.

The above changes should have no effect on the output. However I then
added state tracking for whether the scratch register or the save slot
already contain the right value and therefore don't need to be the
target of another move. This should remove the need for the
redundant-move eliminator to clean up after parallel moves.

My first approach assumed the correct value was in exactly one of these
places, but both allocations can be correct, and we can save more copies
by acknowledging that.

* Remove MoveAndScratchResolver::new

This struct now contains exactly the fields whose values are passed to
`new`. I feel it's better API design in this case to use the named
fields than to expect callers to keep track of the positional arguments
to the `new` method.

* Clarify relationship between save_slot/scratch_dirty

* Review comments: factor out is_stack_to_stack_move

* Remove "victim" terminology from public API

Also update more comments I missed.