Move the split point to the first fixed use (#165)
Address https://github.com/bytecodealliance/wasmtime/issues/7147 by
moving the bundle split point to the first fixed use found after the
suggested split point. This implementation only considers the first
liverange that contains the split point, so there's room for
improvement.

The intent with this change is to not artificially constrain the bundle
produced after the split point by leaving unconstrained uses that could
be kept above the split point. Splitting at the first fixed use after
the suggested split point should produce a less constrained upper
bundle, which will be easier to allocate.

One interesting effect from the example in the issue linked above is
that the change of split point alone does not fix the problem, as we
also trim the space around the split point into the spill bundle. Doing
this creates the same splits that we would have arrived at before, as
the second load doesn't use `v1` all and thus that region is trimmed to
the spill bundle. I addressed that by disabling trimming when the split
point is advanced, under the assumption that the point moving means that
you probably want to give the allocator another chance with both
bundles.

I'm not sure that this implementation is the right path forward
currently, as it might be better to perform the heuristic outside of the
call to `split_and_requeue_bundle`, allowing us the flexibility to make
the decision about trimming at that point.