Allow fixed-early-def and fixed-late-use constraints (#168)
These are useful to model fixed registers which should not be reused for
other uses/defs. These were disallowed in #54, but this was too
conservative.

Fundamentally, the only situation where a preg can be used in multiple
fixed constraints within a single instruction is with an early-use and a
late-def. Anything else is a user error because the live ranges will
overlap.

As such this PR relaxes the operand rewrite from #54 to only apply in
this specific situation. Fixed-late-use and fixed-early-def operands are
left unchanged.