Verify that branches with operands are properly supported
This extends the fuzzer and checker to verify that branches that produce
outputs are properly supported and no moves are inserted after a branch.

This also checks the one edge case where a branch instruction may have
both operands and blockparams: when there is only one successor and that
successor only has one predecessor.