Revert "Make parallel-moves resolver available publicly" (#175)
Reverts bytecodealliance/regalloc2#156

We never used this functionality in Cranelift. When we tried, we
discovered that's it's pretty tricky to use correctly outside of the
specific context it was written for in RA2, and then we found out that
we didn't need it at all. So let's revert making it public.