Improve performance by optimizing  bundle merging logic. (#204)
While merging two bundles, the `merge_bundles` function appends two
sorted vectors and sort the resulting vector again. This approach used
to be fast in most cases, but rustc 1.81 introduced changes in sorting
algorithm that made `sort_unstable_by_key` to behave very bad with
vectors that are almost sorted.

This introduces an optimization consisting in handling the special case
where the vector being appended contains a single item differently. This
case is very common, and there's a benefit in handling it differently
both with rust 1.81 and with earlier versions.

Fixes #203