Redefine a minimal bundle to have only one LiveRange. (#214)
When exerting additional pressure on regalloc with
bytecodealliance/wasmtime#10502, which can lead to call instructions
that have significantly more (`any`-constrained) defs, we hit panics in
RA2 where (i) a bundle merged several LiveRanges, (ii) one of these
LiveRanges had a fixed-reg constraint on an early use, (iii) this
fixed-reg constraint conflicted with a clobber (which always happens at
the late-point), (iv) the bundle merged in another LiveRange of some
arbitrary def at the late point.

This would make a bundle (which is the atomic unit of allocation) that
covers the whole inst, including the late point; and is required to be
in the fixed reg; this is unallocatable because the clobber is also at
the late point in that reg.

Our allocate-or-split-and-retry logic does not split if a bundle is
"minimal". This is meant to give a base case to the retries: when
bundles break down into their minimal pieces, any solvable set of
constraints should result in allocations.

Unfortunately the "is minimal" predicate definition did not account for
multiple LiveRanges, but rather only tested whether the total
program-point range of the bundle was over one instruction. If there are
multiple LiveRanges, we can still split them off, and the resulting
split bundles may cover only half the instruction, avoiding the
clobbers.