Fix minimal-bundle splitting to avoid infinite loops.
This PR reworks minimal-bundle logic to be consistent between property
computation and splitting, and overall simpler.

To recap: a "minimal bundle" is an allocation bundle that is as small as
possible. Because RA2 does backtracking, i.e., can kick a bundle out of
an allocation to make way for another, and because it can split, i.e.
create more work for itself, we need to ensure the algorithm "bottoms
out" somewhere to ensure termination. The way we do this is by defining
a "minimal bundle": this is a bundle that cannot be split further. A
minimal bundle is never evicted, and any allocatable input (set of
uses/defs with constraints), when split into minimal bundles, should
result in no conflicts. We thus want to define minimal bundles as
*minimally* as possible so that we have the maximal solving capacity.

For a long time, we defined minimal bundles as those spanning a single
instruction (before- and after- progpoints) -- because we cannot insert
moves in the middle of an instruction. In #214, we updated
minimal-bundle splitting to avoid putting two different unrelated uses
in a single-instruction bundle, beacuse doing so and calling it
"minimal" (unsplittable) can artificially extend the liverange of a use
with a fixed-reg constraint at the before-point into the after-point,
causing an unsolveable conflict. This was triggered by new and tighter
constraints on callsites in Cranelift after
bytecodealliance/wasmtime#10502 (merging retval defs into calls) landed.

Unfortunately this also resulted in an infinite allocation loop, because
the definition of "minimal bundle" did not agree between the
split-into-minimal-bundles fallback/last-ditch code, and the bundle
property computation. The splitter was splitting as far as it was
willing to go, but our predicate didn't consider those bundles minimal,
so we continued to re-attempt splitting indefinitely.

While investigating this, I found that the minimal-bundle concept had
accumulated significant cruft ("the detritus of dead fuzzbugs") and this
tech-debt was making things more confusing than not -- so I started by
clearly defining what a minimal bundle *is*. Precisely:

- A single use, within a single LiveRange;
- With that LiveRange having a program-point span consistent with the
  use:
  - Early def: whole instruction (must live past Late point so it can
    reach its uses; moves not possible within inst);
  - Late def: Late point only;
  - Early use: Early point only;
  - Late use: whole instruction (must be live starting at Early so the
    value can reach this use; moves not possible within inst).

This is made easier and simpler than what we have before largely
because the minimal-bundle splitter aggressively puts spans of LiveRange
without uses into the spill bundle, and because we support overlapping
LiveRanges for a vreg now (thanks Trevor!), so we can rely on having
*some* connectivity between the def and its uses even if we aggressively
trim LiveRanges in the minimal bundles down to just their defs/uses.

The split-at-program-point splitter (i.e., not the fallback
split-into-minimal-bundles splitter) also got a small fix related to
this: it has a mode that was intended to "split off one use" if we enter
with a split-point at the start of the bundle, but this was really
splitting off all uses at the program point (if there are multiple of
the same vreg at the same program point). In the case that we still need
to split these apart, this just falls back to the minimal-bundle
splitter now.

Fixes #218, #219.