API documentation update: add a note about clobber/def conflicts. (#223)
This arose in #222. It has been the case for a while that we do not
permit clobbers and defs to name the same physical register; this is
because clobbers become defs internally. Unfortunately the doc-comment
on the relevant part of the `Function` trait was out-of-date.

Ideally we would also detect this during validation or liverange
construction rather than panic'ing later, but that is a separate concern
(panic-cleanliness).

Note that this doesn't affect Cranelift; this is a concern mainly for
docs correctness and for third-party users.