Checker: process defs/clobbers on branches to allow checking of try-calls.
The regalloc checker previously handled branches specially: it skipped
the normal operand-procesing logic because it handles block-call
arguments explicitly with parallel moves.

However, this is no longer fully correct in the presence of try-calls:
these are branches but also have normal uses, defs, and clobbers, so we
need to process those normally.

The comment noted also that we needed to skip operand processing on
branches because edge-moves may come before branches, but I'm not sure
that exemption ever made sense: if moves interfere with operands to the
branch (e.g. conditional branch inputs) we should catch that too.

This simplifies the checker's instruction handling to no longer skip
branches' operands, with no other changes (and no changes to the
allocator itself).

Addresses bytecodealliance/wasmtime#10585 (and will fix once we release
with this and pull into Cranelift).