Re-add support for stack constraints. (#231)
In bytecodealliance/wasmtime#11285, we are discussing various approaches
to allow generated code to communicate some "dynamic context" from
exception-handling callsites to a stack-walker, without requiring
changes to the overall frame format. The approach that seems to be the
best choice so far is to add a stack-constrained vreg use to a try-call
instruction and then communicate its location on the stack to the
stack-walker in the exception metadata. The upshot of this is that we
need support for stack-constrained operands.

Unfortunately we removed this support in #185 because we had removed the
last (at the time) use-case for it, namely, GC stackmaps. Fortunately,
reverting that PR (and most of a subsequent one, #192, that fixed up the
removal), seems to be fairly straightforward. Those two reverts plus a
few small fixups are included in this PR. It seems to fuzz cleanly when
fuzzed briefly; I will fuzz it overnight and also make use of it in
above-mentioned use-case to ensure it works properly.