Ensure all fixed registers fit within their `Limit`
Operands can be constrained to fixed registers, potentially multiple
_different_ fixed registers; `Env::fixup_multi_fixed_vregs` fixes this.
Here we also check that, for all the uses of a fixed register, the fixed
register chosen does not fall outside of any limit constraints. To do
this, we find the _lowest_ (minimum) limit in all uses of the operand
and ensure the fixed register is below that.