Constrain iteration of registers with a `Limit`
During allocation, we iterate over the available registers and pick one
within all the constraints (i.e., fixed, hints, preferred/non-preferred
lists). This adds a limit to all of those constraints: unless the
register is fixed, we only pick a register with a HW encoding below the
limit, if one is given.