Pass hints to `RegTraversalIter` as `Option<PReg>` (#241)
As noted in a [comment] to #235, passing an `Option<PReg>` _is_ a bit
more clear API. To do this, this change adds `PReg::is_valid()` to do
the conversion. This change (rather unnecessarily) also renames this
variable to `hint` everywhere (previously: `hint`, `hint_reg`,
`reg_hint`) and uses variables directly in several close-by `format!`
strings.

This is a non-functional change.

[comment]:
https://github.com/bytecodealliance/regalloc2/pull/235#pullrequestreview-3236527918