fuzz: improve `Options` control over generated functions
Previously, the ranges of the knobs for generating components of a
functions were hard-coded in `Func::arbitrary_with_options`. This change
extracts some of them to `Options` to (a) allow for more control from
outside and (b) to explain what these parameters are doing.

The one functional change here is that `num_clobbers_per_inst = 0..=10`
is used both times clobbers are created (callsite-ish arm and clobbers
arm); previously the clobbers arm only clobbered `0..=5` registers. It
seemed possible that this was just an oversight (?).