Only cache limits as `u8`
To avoid possible cache issues, Chris [asked] to store a smaller size
for the cached limit. This change drops the `Option<usize>` to a
`Option<u8>` adding conversions between `usize` and `u8` as needed. The
`usize` is retained for consistency in the public API
(`OperandConstraint::Limit`) and for ease of use in comparing with HW
encodings and slice indexes (all `usize`).

[asked]: https://github.com/bytecodealliance/regalloc2/pull/239#discussion_r2361118861