Avoid extra block calculation during fuzz generation
In the process of building up a control-flow graph of blocks within a
generated function, these lines appear to re-calculate the edge links
between successors and predecessors. This _should_ be taken care of by
`FuncBuilder::add_edge` or so it would appear. This change removes the
extra calculation, assuming it was a leftover from some incomplete
refactoring during the initial work on this crate.