Improve the generation of fuzz functions (#243)
This is intended as a non-functional refactor (with one caveat below).
The generation of functions is quite tricky: `regalloc2` must always
generate test case functions that _are_ allocatable. But the complex
logic of `Func::arbitrary_with_options` makes this more difficult than
it already is. This change attempts to simplify, document, and allow
users of the `Func::arbitrary_with_options` more control over the
generated test cases.

The one caveat is the number of clobbers. Previously, we generated
different amounts of clobbers if `opts.fixed_regs &&
opts.callsite_ish_constraints` and if `opts.clobbers`: `0..=10` versus
`0..=5`. I suspect this is an oversight (?) so I just used the same
`opts.num_clobbers_per_inst` in both cases. This does change the fuzzing
DNA, invalidating the corpus, but this will happen soon anyways if we
start fuzzing `OperandConstraint::Limit`.